Put entrypoint in /, use COPY instead of ADD
I'm inclined not to mess with anything in the Flink release directory
if we don't have to. Additionally, there is precedent for putting
`docker-entrypoint.sh` in `/`, including the Docker official images
documentation[1].

ADD has some weird behaviors (such as supporting URLs) while COPY
does only what it says on the tin, which is what we want here.

[1]https://github.com/docker-library/official-images/#consistency