[FLINK-15831] Remove PR check for changed files
General contributions to this repo (e.g. to the Docker entrypoint
script) should not modify the generated Dockerfiles, hence this check.

However, going forward, changes to the Dockerfiles _will_ come via pull
requests, so this restriction should be removed.

Ideally, we could enforce that either the Dockerfiles (and perhaps
`add-version.sh`) OR any other file are changed in any given PR.